Create subdirectory backend and move TaskList, Storage files into it
