+++ src/main/java/duke/backend/TaskList.java
@@ -1,7 +1,8 @@
-package duke;
+package duke.backend;
 
 import java.util.ArrayList;
 
+import duke.DukeException;
 import duke.tasktype.Deadline;
 import duke.tasktype.Event;
 import duke.tasktype.Task;
